Modularize code and rebase
