+++ workflows/train_durations.py
@@ -6,25 +6,16 @@ from pathlib import Path
 import librosa
 import luigi
 import pandas as pd
-from luigi.contrib.external_program import ExternalProgramTask
-from luigi.contrib.gcs import GCSTarget
-from luigi.parameter import ParameterVisibility
 from tqdm import tqdm
+from workflow.utils.rsync import GSUtilRsyncTask
+
+from workflows.utils.mixin import DynamicRequiresMixin
 
 
 def read_path(path: Path) -> dict:
     y, sr = librosa.load(path)
     duration = librosa.get_duration(y=y, sr=sr)
     return dict(filename="/".join(path.parts[-2:]), duration=duration)
-
-
-class DynamicRequiresMixin:
-    dynamic_requires = luigi.Parameter(
-        default=[], visibility=ParameterVisibility.HIDDEN
-    )
-
-    def requires(self):
-        return self.dynamic_requires
 
 
 class TrainDurations(luigi.Task, DynamicRequiresMixin):
@@ -51,29 +42,6 @@ class TrainDurations(luigi.Task, DynamicRequiresMixin):
         # ensure the output directory exists
         Path(self.output_path).parent.mkdir(parents=True, exist_ok=True)
         df.to_parquet(self.output_path)
-
-
-class GSUtilRsyncTask(ExternalProgramTask, DynamicRequiresMixin):
-    input_path = luigi.Parameter()
-    output_path = luigi.Parameter()
-    is_dir = luigi.BoolParameter(default=True)
-    capture_output = True
-
-    def program_args(self):
-        in_path = self.input_path.rstrip("/")
-        out_path = self.output_path.rstrip("/")
-
-        if self.is_dir:
-            return f"gsutil -m rsync -r {in_path}/ {out_path}/".split()
-        else:
-            return f"gsutil -m rsync -r {in_path} {out_path}".split()
-
-    def output(self):
-        is_gs_path = self.output_path.startswith("gs://")
-        if is_gs_path:
-            return GCSTarget(self.output_path)
-        else:
-            return luigi.LocalTarget(self.output_path)
 
 
 class TrainDurationsWorkflow(luigi.WrapperTask):
